requirements.txt
flask
yfinance
pandas
numpy
matplotlib
mplfinance
scipy
TA-Lib
langchain
langchain-openai
langchain-anthropic
langchain-qwq
langchain-core
langgraph
openai
anthropic
ipython
Pillow
requests
github-copilot-sdk
